cleaned up zshrc a bit

+++ .zshrc
@@ -1,7 +1,25 @@
+SYSINFO="$(uname -s)"
+case "${SYSINFO}" in
+    Linux*)     export CORE_OS=Linux;;
+    Darwin*)    export CORE_OS=MacOS;;
+    *)          export CORE_OS="UNKNOWN:${SYSINFO}"
+esac
+
+unsetopt inc_append_history
+unsetopt share_history
+export GOPATH=$HOME/go
+export PATH=$PATH:$GOPATH/bin:/opt/homebrew/bin
+export TERRAGRUNT_PARALLELISM=1
 export KUBECONFIG=~/.kube/config:~/.kube/bootstrap.conf:~/.kube/xsoar8.conf
 export ZSH_DISABLE_COMPFIX=true
-export ZSH=$HOME/.oh-my-zsh
-ZSH_THEME="af-magic"
+
+# SSH Agent for 1Password and MacOS
+if [[ $CORE_OS == "MacOS" ]]; then
+    AGENT="$HOME/.1password/agent.sock"
+    if [ -S $AGENT ]; then
+        export SSH_AUTH_SOCK=$AGENT
+    fi
+fi
 
 plugins=(
   git
@@ -10,26 +28,18 @@ plugins=(
   kube-ps1
 )
 
-#source $ZSH/oh-my-zsh.sh
-unsetopt inc_append_history
-unsetopt share_history
-export GOPATH=$HOME/go
-export PATH=$PATH:$GOPATH/bin:/opt/homebrew/bin
-export TERRAGRUNT_PARALLELISM=1
-
-# SSH Agent for 1Password and MacOS
-AGENT="$HOME/.1password/agent.sock"
-if [ -S $AGENT ]; then
-    export SSH_AUTH_SOCK=$AGENT
-fi
-
+# Aliases
 alias k=kubectl
 alias ds="docker run --rm -i -t --entrypoint=/bin/bash"  
 alias dssh="docker run --rm -i -t --entrypoint=/bin/sh"
 alias xs=demisto-sdk
-alias ztheme='(){ export ZSH_THEME="$@" && source $ZSH/oh-my-zsh.sh }'
 alias shpod="k attach -n shpod -ti shpod"
+
+# Legacy items to remove
+alias ztheme='(){ export ZSH_THEME="$@" && source $ZSH/oh-my-zsh.sh }'
 alias ss='eval "$(starship init zsh)"'
+export ZSH=$HOME/.oh-my-zsh
+ZSH_THEME="af-magic"
 
 function dsh() {  
     dirname=${PWD##*/}
@@ -42,7 +52,12 @@ function dsshh() {
 }
 
 function clear_cache() {
-    sudo dscacheutil -flushcache; sudo killall -HUP mDNSResponder
+    if [[ $CORE_OS == "MacOS" ]]; then
+        sudo dscacheutil -flushcache; sudo killall -HUP mDNSResponder
+        echo "Cache cleared."
+    else
+        echo "Unsupported OS for this command. Please add support for $CORE_OS."
+    fi
 }
 
 function cortex_env() {
@@ -51,8 +66,10 @@ function cortex_env() {
     TEMP_AUTH_ID=$(op read op://private/$1/username)
     if [ -z "${TEMP_AUTH_ID}" ]; then
         unset XSIAM_AUTH_ID
+        echo "Loaded $1 environment for XSOAR."
     else
         export XSIAM_AUTH_ID=$TEMP_AUTH_ID
+        echo "Loaded $1 environment for XSIAM."
     fi
 }
